add guest export macro
