+++ backends/aws-lambda/iomod/plugins/dynamodb/guest/Cargo.toml
@@ -13,6 +13,5 @@ bytes = "0.5.4"
 base64 = "0.12.1"
 paste = "0.1.12"
 
-assemblylift_core_guest = { version = "0.1.0", package = "assemblylift-core-guest", path = "../../../../../../core/guest" }
 assemblylift_core_event_guest = { version = "0.1.1", package = "assemblylift-core-event-guest", path = "../../../../../../core/event/guest" }
 assemblylift_core_iomod_guest = { version = "0.1.0", package = "assemblylift-core-iomod-guest", path = "../../../../../../core/iomod/guest" }
